+++ lib/zaikio/oauth_client/test_helper.rb
@@ -3,6 +3,23 @@ module Zaikio
     module TestHelper
       extend ActiveSupport::Concern
 
+      VirtualAccessToken = Struct.new(:bearer_id, :bearer_type, :audience, :expired?, keyword_init: true)
+
+      class << self
+        def find_active_access_token(id)
+          return unless id.to_s.starts_with?("AT:")
+
+          _, audience, person_id = id.split(":")
+
+          VirtualAccessToken.new(
+            bearer_id: person_id,
+            audience: audience,
+            bearer_type: "Person",
+            expired?: false
+          )
+        end
+      end
+
       class TestSessionController < ActionController::Base # rubocop:disable Rails/ApplicationController
         def show
           if session[params[:key]].nil?
@@ -43,7 +60,15 @@ module Zaikio
         get "/zaikio/oauth_client/test_helper/session", params: { id: value, key: key }
       end
 
-      def logged_in_as(person)
+      def logged_in_as(person, access_token: nil, client_name: nil)
+        client_name ||= Zaikio::OAuthClient.client_name ||
+                        Zaikio::OAuthClient.configuration.all_client_names.first
+        set_session(
+          :zaikio_access_token_id,
+          access_token&.id || "AT:#{client_name}:#{person.id}"
+        )
+
+        # Deprecated please use zaikio_access_token_id
         set_session(:zaikio_person_id, person.id)
       end
     end